Move branch_trees module to a new create
