+++ Cargo.toml
@@ -36,6 +36,7 @@ members = [
     "crates/gitbutler-stack",
     "crates/gitbutler-hunk-dependency",
     "crates/gitbutler-settings",
+    "crates/gitbutler-workspace",
 ]
 resolver = "2"
 
@@ -95,6 +96,7 @@ gitbutler-stack = { path = "crates/gitbutler-stack" }
 gitbutler-forge = { path = "crates/gitbutler-forge" }
 gitbutler-hunk-dependency = { path = "crates/gitbutler-hunk-dependency" }
 gitbutler-settings = { path = "crates/gitbutler-settings" }
+gitbutler-workspace = { path = "crates/gitbutler-workspace" }
 
 [profile.release]
 codegen-units = 1 # Compile crates one after another so the compiler can optimize better
