+++ crates/gitbutler-branch-actions/Cargo.toml
@@ -30,6 +30,7 @@ gitbutler-cherry-pick.workspace = true
 gitbutler-oxidize.workspace = true
 gitbutler-stack.workspace = true
 gitbutler-hunk-dependency.workspace = true
+gitbutler-workspace.workspace = true
 serde = { workspace = true, features = ["std"] }
 bstr.workspace = true
 diffy = "0.4.0"
@@ -49,6 +50,7 @@ toml.workspace = true
 once_cell = "1.20"
 pretty_assertions = "1.4"
 gitbutler-testsupport.workspace = true
+gitbutler-workspace.workspace = true
 gix = { workspace = true, features = ["max-performance"] }
 gitbutler-git = { workspace = true, features = ["test-askpass-path"] }
 glob = "0.3.1"
