+++ crates/gitbutler-branch-actions/tests/virtual_branches/mod.rs
@@ -57,7 +57,6 @@ impl Test {
 
 mod amend;
 mod apply_virtual_branch;
-mod branch_trees;
 mod create_commit;
 mod create_virtual_branch_from_branch;
 mod init;
