+++ crates/gitbutler-edit-mode/src/lib.rs
@@ -5,9 +5,6 @@ use std::str::FromStr;
 use anyhow::{bail, Context, Result};
 use bstr::ByteSlice;
 use git2::build::CheckoutBuilder;
-use gitbutler_branch_actions::branch_trees::{
-    checkout_branch_trees, compute_updated_branch_head, BranchHeadAndTree,
-};
 use gitbutler_branch_actions::internal::list_virtual_branches;
 use gitbutler_branch_actions::{update_workspace_commit, RemoteBranchFile};
 use gitbutler_cherry_pick::{ConflictedTreeKey, RepositoryExt as _};
@@ -26,6 +23,7 @@ use gitbutler_reference::{ReferenceName, Refname};
 use gitbutler_repo::{rebase::cherry_rebase, RepositoryExt};
 use gitbutler_repo::{signature, SignaturePurpose};
 use gitbutler_stack::{Stack, VirtualBranchesHandle};
+use gitbutler_workspace::{checkout_branch_trees, compute_updated_branch_head, BranchHeadAndTree};
 use serde::Serialize;
 
 pub mod commands;
