+++ crates/gitbutler-workspace/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "gitbutler-workspace"
+version = "0.0.0"
+edition = "2021"
+authors = ["GitButler <gitbutler@gitbutler.com>"]
+publish = false
+
+[dependencies]
+anyhow.workspace = true
+tracing.workspace = true
+git2.workspace = true
+gitbutler-stack.workspace = true
+gitbutler-repo.workspace = true
+gitbutler-project.workspace = true
+gitbutler-commit.workspace = true
+gitbutler-oxidize.workspace = true
+gitbutler-command-context.workspace = true
+gitbutler-cherry-pick.workspace = true
+
+
+[[test]]
+name = "workspace"
+path = "tests/mod.rs"
+
+[dev-dependencies]
+gitbutler-testsupport.workspace = true
+gitbutler-branch.workspace = true
+gitbutler-branch-actions.workspace = true
+gitbutler-workspace.workspace = true
