+++ crates/gitbutler-workspace/src/branch_trees.rs
@@ -1,4 +1,3 @@
-use crate::VirtualBranchesExt as _;
 use anyhow::{bail, Result};
 use gitbutler_cherry_pick::RepositoryExt;
 use gitbutler_command_context::CommandContext;
@@ -7,7 +6,7 @@ use gitbutler_oxidize::{git2_to_gix_object_id, gix_to_git2_oid};
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_repo::rebase::cherry_rebase_group;
 use gitbutler_repo::{GixRepositoryExt, RepositoryExt as _};
-use gitbutler_stack::Stack;
+use gitbutler_stack::{Stack, VirtualBranchesHandle};
 use tracing::instrument;
 
 /// Checks out the combined trees of all branches in the workspace.
@@ -19,7 +18,7 @@ pub fn checkout_branch_trees<'a>(
     _perm: &mut WorktreeWritePermission,
 ) -> Result<git2::Tree<'a>> {
     let repository = ctx.repository();
-    let vb_state = ctx.project().virtual_branches();
+    let vb_state = VirtualBranchesHandle::new(ctx.project().gb_dir());
     let branches = vb_state.list_branches_in_workspace()?;
 
     if branches.is_empty() {
@@ -147,100 +146,5 @@ pub fn compute_updated_branch_head_for_commits(
             head: new_head,
             tree: rebased_tree.tree_id(),
         })
-    }
-}
-
-// These possibly could be considered more "integration" tests, but there is no
-// need for `checkout_branch_trees` to be public, so it is tested here.
-#[cfg(test)]
-mod test {
-    use std::fs;
-
-    use gitbutler_branch::BranchCreateRequest;
-    use gitbutler_command_context::CommandContext;
-    use gitbutler_repo::RepositoryExt as _;
-    use gitbutler_testsupport::{paths, testing_repository::assert_tree_matches, TestProject};
-
-    #[test]
-    fn checkout_with_two_branches() {
-        let test_project = &TestProject::default();
-
-        let data_dir = paths::data_dir();
-        let projects = gitbutler_project::Controller::from_path(data_dir.path());
-
-        let project = projects
-            .add(test_project.path())
-            .expect("failed to add project");
-
-        crate::set_base_branch(&project, &"refs/remotes/origin/master".parse().unwrap()).unwrap();
-
-        let branch_1 =
-            crate::create_virtual_branch(&project, &BranchCreateRequest::default()).unwrap();
-
-        fs::write(test_project.path().join("foo.txt"), "content").unwrap();
-
-        crate::create_commit(&project, branch_1, "commit one", None, false).unwrap();
-
-        let branch_2 =
-            crate::create_virtual_branch(&project, &BranchCreateRequest::default()).unwrap();
-
-        fs::write(test_project.path().join("bar.txt"), "content").unwrap();
-
-        crate::create_commit(&project, branch_2, "commit two", None, false).unwrap();
-
-        let tree = test_project.local_repository.create_wd_tree().unwrap();
-
-        // Assert original state
-        assert_tree_matches(
-            &test_project.local_repository,
-            &tree,
-            &[("foo.txt", b"content"), ("bar.txt", b"content")],
-        );
-        assert_eq!(tree.len(), 2);
-
-        // Checkout an empty tree
-        {
-            let tree_oid = test_project
-                .local_repository
-                .treebuilder(None)
-                .unwrap()
-                .write()
-                .unwrap();
-            let tree = test_project.local_repository.find_tree(tree_oid).unwrap();
-            test_project
-                .local_repository
-                .checkout_tree_builder(&tree)
-                .force()
-                .remove_untracked()
-                .checkout()
-                .unwrap();
-        }
-
-        // Assert tree is indeed empty
-        {
-            let tree: git2::Tree = test_project.local_repository.create_wd_tree().unwrap();
-
-            // Tree should be empty
-            assert_eq!(
-                tree.len(),
-                0,
-                "Should be empty after checking out an empty tree"
-            );
-        }
-
-        let ctx = CommandContext::open(&project).unwrap();
-        let mut guard = project.exclusive_worktree_access();
-
-        super::checkout_branch_trees(&ctx, guard.write_permission()).unwrap();
-
-        let tree = test_project.local_repository.create_wd_tree().unwrap();
-
-        // Should be back to original state
-        assert_tree_matches(
-            &test_project.local_repository,
-            &tree,
-            &[("foo.txt", b"content"), ("bar.txt", b"content")],
-        );
-        assert_eq!(tree.len(), 2, "Should match original state");
     }
 }
