+++ crates/gitbutler-workspace/src/lib.rs
@@ -1,0 +1,6 @@
+mod branch_trees;
+
+pub use branch_trees::{
+    checkout_branch_trees, compute_updated_branch_head, compute_updated_branch_head_for_commits,
+    BranchHeadAndTree,
+};
